+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added `pageSize` field to `VisitaConnection` type in `schema.py` to indicate the number of items requested per page.
 
 ### Changed